+++ test-java/orchard/java/PrivateFieldClass.java
@@ -1,0 +1,11 @@
+package orchard.java;
+
+public class PrivateFieldClass {
+
+    public PrivateFieldClass(int age){
+        this.age = age;
+    }
+
+    private final int age;
+
+}